no-op: move consts to their own file
